Added a menu and functions.
Added a function to calculate the cost-basis for each trade with a GBP trade pair. This is nested inside the separate_coins function and which is now part of the “Upload information to new files” menu option.

Created “alternative_pairs” function to store logic from the main code. This writes any trades that are without a calculable cost-basis.